Separate object-oriented style and chaining in docs
I think the fact that the object-oriented style and chaining share
implementation details has lead us to conflate them in the docs. In practice,
the two features have almost nothing to do with each other.

The existing documentation could easily lead one to believe that they could
start a chain using the OOP style.

This confusion is compounded by a few factors:

1. Lodash supports chains starting with the OOP wrapper.
2. The fact that we share names with native Array methods can lead to manual
   tests that appear to work.

For example:

    _([1,2,3]).reverse().map(function(n){ return n * 2; });

This appears to be an example of successfully starting a chain with the OOP
style. The astute observer will note that the lack of a `.value()` call, proves
this is not Underscore chaining, but a new user attempting to clarify the
behavior for herself may not catch that detail.