Add `npm run weight` tool
The way a code change maps to a change in minified/gzipped size is not always
intuitive (#2383). To help folks in our effort to slim down (#2060), they can
now issue the command:

    npm run weight

And get output like:

    6.27 kB

Which should give a fairly good estimate of the "over-the-wire" size of their
`underscore.js`.

Caveat: The weight reported does not include the source-map comment at the end
of the file.