Add 4ms slack to _.throttle tests sensitive to setTimeout imprecision
These tests fail relatively often, randomly, though more in some
engines than in others. The random failures are a combination of a
race condition and the imprecision of the global setTimeout function.

This measure shouldn't be taken lightly. It can only be applied if it
does not affect the timing of the throttled (or debounced) function
invocations. In the current patch, only the time at which the result
is read out is slightly delayed so that the readout is less likely to
be too early.

Part of #2879, also related to #2802.